Unit tests refactor
- Tests are in two different directories depending on the profile
  server they test on (fake or real)
- Created RealProfileServerTestCase base class which skip tests
  if the configuration to connect to the real profile server is not
  in place
